Update patch set 7
Patch Set 7:

As mentioned in previous comments, our problem with this is that the scenario involves sending transactions from the BBinder destructor (step #5). We haven't closed the loop yet with our frameworks team on whether this is supported (some of the key folks are just now getting back from their holiday).

The change itself is probably OK, but we'd like the rationale to be based on a real scenario (performance or behavior in a supported configuration). As Arve points out in his comment above, this change might fix your particular case, but may not be a complete fix for the general case of transactions during destructors.

So, if this is not a supported scenario, we probably won't take this change (at least not based on this rationale). If it is supported, we would probably need a more complete solution (but in this case your change might be a reasonable first step toward the complete solution).

Patch-set: 7