Update patch set 7
Patch Set 7:

Ok, after talking to the frameworks team, sending transactions during the destructor is considered bad practice and can result in the kind of issue you observed (and as discussed, the fix you propose is not a complete solution if this were to be a fully supported operation).

Is there a reason why your userspace program cannot be changed to eliminate item #5?

In your commit message, you also mention load balancing and parallelism benefits. Do you have any workload that has tangible benefit from this? If this has an actual performance benefit, we can consider it.

If you can quantify the performance/scalability benefit, please post the data on the android-4.4 branch (not here in android-4.4.y).

Patch-set: 7
Reviewer: Todd Kjos <1069543@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0