Update patch set 1
Patch Set 1:

Ok. I think what you're saying is its perfectly Ok to have a migrated fake cfs task to be in the original rt_rq's cfs_throttled list. But one fact is that after the task is promoted back to fifo, it will be added to the rt_rq of the CPU that the fake-CFS task is currently running on, which will have a different rt_rq than the original one. Maybe that's not a problem?

The promotion in setprio_fifo happens on the current task CPU:
 enqueue_task(cpu_rq(cpu), p, ENQUEUE_REPLENISH);

Ignore my noise if that's a perfectly Ok scenario :)

Patch-set: 1