Update patch set 1
Patch Set 1:

when the task is promoted back to fifo it gets added to the original rt_rq.  cfs_unthrottle_rt_tasks is called on a rt_rq that contains a list of all fake CFS tasks that may be on other CPUs, so the promotion is initiated from the original rt_rq.  This is why we lock the other rq in __setprio_fifo.  No worries, the discussion is good and can help find any issues!

Patch-set: 1