dma-buf/sw_sync: Partially restore in-kernel API.
When the sw-sync driver was moved into drivers/dma-buf, its in-kernel API
was removed. Some Android drivers such as goldfish and ADF are still
using this API, so they no longer build. Bring back a small subset of
the API for these drivers.

ADF will likely be superseded by explicit fencing for atomic DRM drivers,
but this feature did not make it into Linux 4.9, so ADF still has
superior functionality to DRM, and it needs to work.

Change-Id: I66ef949513745e660c6431ccdb15a8bdfdc40979
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>