oom: Add notification for oom_score_adj
This adds subscribtion for changes in oom_score_adj, this
value is important to android systems.
For task that uses oom_score_adj they read the task list.
This can be long and need rcu locks and has a impact
on the system. Let the user track the changes based on
oom_score_adj changes and keep them in their own context
so they do their actions with minimal system impact.

Change-Id: Ia802ba749a813ea6223bd1522f2041c8bd6e003b
Signed-off-by: Peter Enderborg <peter.enderborg@sonymobile.com>