mm: Remove RCU and tasklocks from lmk
Fundamental changes:
1 Does NOT take any RCU lock in shrinker functions.
2 It returns same result for scan and counts.
3 It does not lock any other process than the one that is
going to be killed.

Background.
The low memory killer scans for process that can be killed
to free memory. This can be cpu consuming when there is
a high demand for memory. This can be seen by analysing the
kswapd0 task work.

This patch create a structure within the lowmemory killer
that caches the user spaces process that it might can
kill. It is done with a sorted rbtree so we can very easy find
the candidate to be killed, and knows its properies as
memory usage and oom_score_adj. To be able to achive this it
uses oom_score_notify events.

Test scenario.
Put the device in a state where there is a high demand for
memmory. This is done by allocating memory until there is a
incresead cpu activity from kswapd0. Observe how much ram is
needed to the cpu activity to about 15-20 %. (use top in a shell)

Apply the patch and test with the same amount of allocated ram.

Observe the lower cpu usage in kswapd0.

This test is a very syntetic senario, but this is what is happing
on our devices, but in normal cases the low memmory killer cleans
the memory after a few seconds and phone behave normally after that.

This patch also as a other effect, we are now free to do
other lowmemorykiller configurations.  Without the patch
there is a need for a tradeoff between freed memory and task and rcu
locks. This is no longer a concern for tuning lmk. This patch is not
intended to do any calculation changes other than we do use the cache
for calculate the count values and that makes kswapd0 to shrink other
areas.

Change-Id: I0222cd2ffa4974f8da40ca0ec3a24c1380bea910
Signed-off-by: Peter Enderborg <peter.enderborg@sonymobile.com>