Update patch set 4
Patch Set 4: Presubmit-Ready+1

Changed in v4 of patchset: update commit subjects to be tagged appropriately for the Android kernel, and add or update explanations (in commit messages or comments) of a few things, like the choice of the mode number and the decision to handle swapping coefficients in the HEH template rather than in the poly_hash algorithms.  No functional changes in any of the patches.

Patch-set: 4