Update patch set 3
Patch Set 3:

This is deeper than just checking for bh[b] == NULL in squashfs_bh_to_actor(). The design (for the uncompressed read case, whicih is what triggered the crash) assumes that for every page present in the actor->page[] array in that function, a corresponding bh[b] block would have been read in. The fact that bh[b] is NULL points at a bug somewhere else.

I do find something strange (that I can't explain) in bh_is_optional(), which *might* explain why bh[b] is NULL when we don't expect it to be.

static int bh_is_optional(struct squashfs_read_request *req, int idx)
{
...
...
        if (end_idx >= req->output->pages)
                end_idx = start_idx;
        /*                                                                                       
         * XXX - MOHAN question.                                                                 
         * if the block is uncompressed, if the start_idx page is absent,                        
         * and the NULL idx page is absent, then the assumption is that                          
         * the underlying uncompressed block does not have to be read in                         
         * at all.                                                                               
         * Is this valid ? Could the VFS ask for pages that are in-between ?                     
         * Shouldn't we be checking for the presence of pages in the                             
         * range start_idx->end_idx before deciding if we can skip reading                       
         * the entire uncompressed block ?                                                       
         *                                                                                       
         * Conversely : Why do we not do the same check for compressed blocks                    
         * too ? If we have a compressed block where page[start_idx->end_idx] is                 
         * all NULL, we don't have to read the compressed block either ?                         
         *                                                                                       
         */
        return !req->output->page[start_idx] && !req->output->page[end_idx];  

In general, a squashFS block is 128KB, holding 32 pages. If is possible that the readahead code asks for some number of pages that happen to all fall in a squashFS block but either the first and last page are already resident in the page cache (so the page[] entries are NULL). In which case bh_is_optional() will tell actor_getblks() to not read in that uncompressed block and will NULL out bh[b]. But there could be pages mapping into that block that are in the middle leading to the scenario that led to the crash.

How reproducible is this ? One thing to try is to do the thing I suggest - instead of checking to see if the first and last page are present, check to see if all pages (for that uncompressed block) are present in the page[] array. Even none of the pages for the array are present, only then do we skip reading that block,

Patch-set: 3
Reviewer: Mohan Srinivasan <1105434@85c56323-6fa9-3386-8a01-6480fb634889>