Update patch set 3
Patch Set 3:

Continuing the read the relevant code very carefully, I *think* there is a bug in this piece of logic. Hard for me to say whether this can cause the unexpected (bh[b] == NULL) in the bh to actor code. 

squashfs_bio_submit()
...
...
        /* Create and submit the BIOs */
	for (b = 0; b < nr_buffers; ++b, offset += blksz) {
                bh = req->bh[b];
                if (!bh || !trylock_buffer(bh))
                        continue;
		if (buffer_uptodate(bh)) {
                        unlock_buffer(bh);
                        continue;
                }
If we find there is no bh for this offset, or if the buffer
is uptodate at this offset, we are going to skip reading the
underlying block. Just continue'ing here is a mistake (I think).
Before continuing, we need to check if we have an accumulated
bio already (check for bio != NULL), if there is, we need to 
submit_bio() the accumulated bio, set bio = NULL, then continue.

Patch-set: 3
Reviewer: Mohan Srinivasan <1105434@85c56323-6fa9-3386-8a01-6480fb634889>