ANDROID: binder: fix thread hangs waiting for proc work
With fine-grained locking, there is an uncommon case where
a thread can be waiting on the per-proc wait queue while it
has per-thread work to do. If there isn't any proc work
to do, it may never be awakened so the binder transaction
is hung. This CL fixes the issue by checking if the thread
is waiting for the per-proc queue when waking a thread for
per-thread work.

There is another case fixed here that can also delay this
work in which no thread is awakened from the per-proc queue
because another thread is servicing the queue. If that
thread ends up doing per-thread work, it now wakes a thread
waiting on the per-proc queue.

Change-Id: I18a0fc25abd97ad2eaf704630d1e6257a53f2215
Signed-off-by: Todd Kjos <tkjos@google.com>