Update patch set 1
Patch Set 1:

If the kernel is 64 bit and ANDROID_BINDER_IPC_32BIT=N this should not change anything on the kernel size, since there is already implicit padding there. The padding is needed in this case (on the user-space side). It is not strictly needed when ANDROID_BINDER_IPC_32BIT=Y however and it changes the layout in that case. If anyone has shipped that configuration, we will need a pad-if-ANDROID_BINDER_IPC_32BIT=N macro.

Patch-set: 1
Reviewer: Arve Hjønnevåg <1000205@85c56323-6fa9-3386-8a01-6480fb634889>