Update patch set 1
Patch Set 1:

> If the kernel is 64 bit and ANDROID_BINDER_IPC_32BIT=N this should
 > not change anything on the kernel size, since there is already
 > implicit padding there. The padding is needed in this case (on the
 > user-space side). It is not strictly needed when ANDROID_BINDER_IPC_32BIT=Y
 > however and it changes the layout in that case. If anyone has
 > shipped that configuration, we will need a pad-if-ANDROID_BINDER_IPC_32BIT=N
 > macro.

The only users of this object are hwbinder clients, which will only ship with our next big release. This is also why I'd asked to not upstream this yet - as far as Android is concerned, this interface is only frozen with the release of our next desert.

Patch-set: 1
Reviewer: Martijn Coenen <1013030@85c56323-6fa9-3386-8a01-6480fb634889>