Update patch set 3
Patch Set 3:

> Anyway, the patch should be split up at the least, don't move debugging stuff around and do major lock reworks at the same time :)

Im trying to spend time on details until we can see a future of this set. However it quite funny that people are so picky on the fix that solve things, but very little on the scary stuff like the oom_score_adj notifications that hooks in to fork.c. The other stuff is staging and really need some changes.

Patch-set: 3
Reviewer: Peter Enderborg <1070654@85c56323-6fa9-3386-8a01-6480fb634889>