Update patch set 2
Patch Set 2:

> @GregKH, despite it is probably not easy repro case when resume
 > handler is called in the middle of destroy one,  Subash has a
 > point. Current patch is upstream patch in 4.9. Should we get fix as
 > stable?
 > In 4.10 there is no pm_notifier at all, so there is no upstream fix
 > case.

What is the git commit id of this upstream?

And if you want a commit in the stable trees, just email stable@vger.kernel.org with the git commit id so that it can go into all of the relevant stable releases, not just the ones I maintain.

Patch-set: 2
Reviewer: Greg Kroah-Hartman <1125653@85c56323-6fa9-3386-8a01-6480fb634889>